feat: add javadoc
