+++ ewm-main/src/main/java/ru/practicum/ewmmain/repository/event/EventRepository.java
@@ -43,7 +43,6 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
                                @Param("rangeStart") LocalDateTime rangeStart,
                                @Param("rangeEnd") LocalDateTime rangeEnd, Pageable pageable);
 
-
     /**
      * Поиск только опубликованных событий с учетом фильтров.
      * @param text поиск текста по полям annotation и description класса {@link Event}.
@@ -54,7 +53,6 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
      * @param pageable пагинация.
      * @return возвращает подборку событий в соответствии с заданными фильтрами {@link List}<{@link Event}>.
      */
-    //Только опубликованные события.
     @Query(value = "SELECT e FROM Event e " +
             "WHERE ((:text IS NULL OR lower(e.annotation) LIKE concat('%', :text, '%')) " +
             "OR (:text IS NULL OR lower(e.description) LIKE concat('%', :text, '%'))) " +
@@ -66,9 +64,20 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
                                    @Param("paid") Boolean paid, @Param("rangeStart") LocalDateTime rangeStart,
                                    @Param("rangeEnd") LocalDateTime rangeEnd, Pageable pageable);
 
-
-
-
+    /**
+     * Поиск доступных событий в локации {@link ru.practicum.ewmmain.setlocation.model.SetLocation}.
+     * @param state статус события: PENDING, PUBLISHED, CANCELED. Если null, то для любых статусов.
+     * @param text текст для поиска в содержимом аннотации и описании события. Поиск без учета регистра.
+     * @param categories список id категорий в которых будет вестись поиск. Если null, то по всем категориям.
+     * @param paid поиск только платных/бесплатных событий. Если null, то поиск без фильтра по полю paid.
+     * @param rangeStart дата и время не раньше которых должно произойти событие.
+     * @param rangeEnd дата и время не позже которых должно произойти событие.
+     * @param latitude широта.
+     * @param longitude долгота.
+     * @param radius радиус локации.
+     * @param pageable пагинация.
+     * @return возвращает список событий {@link List}<{@link Event}> с учетом фильтров.
+     */
     @Query(value = "SELECT e FROM Event e " +
             "WHERE (:text IS NULL OR lower(e.annotation) LIKE concat('%', :text, '%')) " +
             "AND (coalesce(:categories, null) IS NULL OR e.category.id IN :categories) " +
@@ -90,6 +99,20 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
                                     @Param("long") Float longitude, @Param("radius") Float radius,
                                     Pageable pageable);
 
+    /**
+     * Поиск событий в локации {@link ru.practicum.ewmmain.setlocation.model.SetLocation}.
+     * @param state статус события: PENDING, PUBLISHED, CANCELED. Если null, то для любых статусов.
+     * @param text текст для поиска в содержимом аннотации и описании события. Поиск без учета регистра.
+     * @param categories список id категорий в которых будет вестись поиск. Если null, то по всем категориям.
+     * @param paid поиск только платных/бесплатных событий. Если null, то поиск без фильтра по полю paid.
+     * @param rangeStart дата и время не раньше которых должно произойти событие.
+     * @param rangeEnd дата и время не позже которых должно произойти событие.
+     * @param latitude широта.
+     * @param longitude долгота.
+     * @param radius радиус локации.
+     * @param pageable пагинация.
+     * @return возвращает список событий {@link List}<{@link Event}> с учетом фильтров.
+     */
     @Query(value = "SELECT e FROM Event e " +
             "WHERE (:text IS NULL OR lower(e.annotation) LIKE concat('%', :text, '%')) " +
             "AND (coalesce(:categories, null) IS NULL OR e.category.id IN :categories) " +
