+++ ewm-main/src/main/java/ru/practicum/ewmmain/setlocation/exception/SetLocationNotFoundException.java
@@ -1,5 +1,11 @@
 package ru.practicum.ewmmain.setlocation.exception;
 
+/**
+ * Локация не найдена.
+ * Extends RuntimeException.
+ *
+ * @author Evgeny S
+ */
 public class SetLocationNotFoundException extends RuntimeException{
     public SetLocationNotFoundException(String message) {
         super(message);
