+++ ewm-main/src/test/java/ru/practicum/ewmmain/setlocation/service/SetLocationServiceImplTest.java
@@ -8,20 +8,20 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
-import ru.practicum.ewmmain.client.StatisticsClient;
-import ru.practicum.ewmmain.event.controller.SortOption;
-import ru.practicum.ewmmain.event.model.category.Category;
-import ru.practicum.ewmmain.event.model.event.Event;
-import ru.practicum.ewmmain.event.model.event.EventDtoMapper;
-import ru.practicum.ewmmain.event.model.event.State;
-import ru.practicum.ewmmain.event.repository.EventRepository;
-import ru.practicum.ewmmain.participationrequest.service.ParticipationReqService;
+import ru.practicum.ewmmain.controller.client.event.SortOption;
+import ru.practicum.ewmmain.model.event.Event;
+import ru.practicum.ewmmain.model.event.State;
+import ru.practicum.ewmmain.model.event.category.Category;
+import ru.practicum.ewmmain.model.user.User;
+import ru.practicum.ewmmain.repository.event.EventRepository;
+import ru.practicum.ewmmain.service.participationrequest.ParticipationReqService;
 import ru.practicum.ewmmain.setlocation.exception.SetLocationNotFoundException;
 import ru.practicum.ewmmain.setlocation.model.SetLocDtoMapper;
 import ru.practicum.ewmmain.setlocation.model.SetLocation;
 import ru.practicum.ewmmain.setlocation.model.SetLocationDto;
 import ru.practicum.ewmmain.setlocation.repository.SetLocationRepository;
-import ru.practicum.ewmmain.user.model.User;
+import ru.practicum.ewmmain.utils.client.StatisticsClient;
+import ru.practicum.ewmmain.utils.mapper.EventDtoMapper;
 
 import java.util.List;
 import java.util.Optional;
